fix(cohere): restore quantized inference and consolidate tests
